python-app/requirements.txt
blinker==1.7.0
click==8.1.7
docopt==0.6.2
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
stomp.py==8.1.0
websocket-client==1.7.0
Werkzeug==3.0.2
